Refactor to show 2 breadcrumbs with different backgrounds on assignment page
