+++ src/containers/pages/PlanAssignment/helpers/JurisdictionTableView/index.tsx
@@ -1,13 +1,10 @@
 import React, { Fragment } from 'react';
 import { Helmet } from 'react-helmet';
 import { RouteComponentProps } from 'react-router';
-import { Breadcrumb, BreadcrumbItem } from 'reactstrap';
-import { LinkList } from '../../../../../components/page/HeaderBreadcrumb/helpers';
+import HeaderBreadcrumb from '../../../../../components/page/HeaderBreadcrumb/HeaderBreadcrumb';
 import { defaultWalkerProps, WithWalkerProps } from '../../../../../components/TreeWalker';
-import { ASSIGN_PLANS, HOME } from '../../../../../configs/lang';
 import { PlanDefinition } from '../../../../../configs/settings';
-import { ASSIGN_PLAN_URL, HOME_URL } from '../../../../../constants';
-import { TreeNode } from '../../../../../store/ducks/opensrp/hierarchies/types';
+import { pagesBuilder } from './helpers/utils';
 
 /** base props for JurisdictionTableView */
 export interface BaseJurisdictionTablProps extends WithWalkerProps {
@@ -35,64 +32,20 @@ export type JurisdictionTableViewProps = RouteComponentProps<RouteParams> &
  */
 const JurisdictionTableView = (props: JurisdictionTableViewProps) => {
   const { plan } = props;
-  const hierarchy = props.hierarchy as TreeNode[];
 
   if (!plan) {
     return null;
   }
 
-  const pageTitle = plan.title;
-  const baseUrl = `${ASSIGN_PLAN_URL}/${plan.identifier}`;
-
-  const initialCurrentPage = {
-    label: pageTitle,
-    url: baseUrl,
-  };
-
-  let currentPage = initialCurrentPage;
-  const pages = [
-    {
-      label: HOME,
-      url: HOME_URL,
-    },
-    {
-      label: ASSIGN_PLANS,
-      url: ASSIGN_PLAN_URL,
-    },
-  ];
-
-  // create breadcrumb props.
-  if (props.match.params.jurisdictionId) {
-    const path = [...hierarchy];
-    const lastNode = path.pop();
-
-    pages.push(initialCurrentPage);
-
-    path.forEach(nd => {
-      pages.push({
-        label: nd.model.label,
-        url: `${baseUrl}/${nd.model.id}`,
-      });
-    });
-
-    currentPage = {
-      label: (lastNode as TreeNode).model.label,
-      url: `${baseUrl}/${(lastNode as TreeNode).model.id}`,
-    };
-  }
+  const breadCrumbProps = pagesBuilder(props);
 
   return (
     <Fragment>
       <Helmet>
-        <title>{pageTitle}</title>
+        <title>{plan.title}</title>
       </Helmet>
-      <div>
-        <Breadcrumb className="reveal-breadcrumb plans-breadcrumb">
-          <LinkList pages={pages} />
-          <BreadcrumbItem active={true}>{currentPage.label}</BreadcrumbItem>
-        </Breadcrumb>
-      </div>
-      <h3 className="mb-3 page-title">{pageTitle}</h3>
+      <HeaderBreadcrumb {...breadCrumbProps} />
+      <h3 className="mb-3 page-title">{plan.title}</h3>
     </Fragment>
   );
 };
